tested accepting and rejecting draw request
